Add an int cast

+++ analysis/src/HeCell.cpp
@@ -98,7 +98,7 @@ int HeCell::fitExponential(double& A0,double& A0s, double& alpha,double& alphas)
    gsl_matrix *covar = gsl_matrix_alloc (p, p);
 
    // This is OK since std::vector is garanteed contiguous
-   struct data d = { n, &_pvst->getX()[0],&_pvst->getY()[0],&_pvst->getE()[0]};
+   struct data d = {(int)n, &_pvst->getX()[0],&_pvst->getY()[0],&_pvst->getE()[0]};
    gsl_multifit_function_fdf f;
    double param_init[2] = { 1.0, 0.0 };
    gsl_vector_view param = gsl_vector_view_array (param_init, p);
